Restore merged Selection interface
